Update file locations of test files to match root structure
